comments, small refactoring
