+++ install.sh
@@ -4,8 +4,9 @@ echo "FMRADSTR: Trying to install apt dependencies..."
 sudo apt install ffmpeg libsndfile1-dev libsoxr-dev git build-essential python3-pip sox-fmt-all && echo "FMRADSTR: Successfully installed apt dependencies!" || echo "FMRADSTR: Failed to install apt dependencies!"
 echo "FMRADSTR: Trying to install/upgrade youtube-dl..."
 sudo sudo pip install --upgrade youtube_dl && echo "FMRADSTR: Successfully installed youtube-dl!" || echo "FMRADSTR: Failed to install youtube-dl!"
-echo "FMRADSTR: Creating core directory..."
+echo "FMRADSTR: Creating directories..."
 mkdir "core"
+mkdir "music"
 echo "FMRADSTR: Downloading PiFmAdv..."
 git submodule update --init --recursive && echo "FMRADSTR: Done!" || echo "FMRADSTR: Not a git repository! Downloading directly..." && git clone https://github.com/miegl/PiFmAdv.git && git checkout 7562fe79789ec4cc0d209a7b14b07d79cdf6e310
 echo "FMRADSTR: Moving and compiling PiFmAdv..."
