+++ pkg/buttons/buttons.go
@@ -19,7 +19,7 @@ var (
 	buttonInvert     = rpio.Pin(12)
 )
 
-// Init sets up four physical buttons.
+// Init sets up physical buttons.
 func Init() {
 	buttons := [4]rpio.Pin{buttonDown, buttonUp, buttonMultiplier, buttonInvert}
 	for _, item := range buttons {
@@ -29,7 +29,7 @@ func Init() {
 	}
 }
 
-// Listen enables frequency, multipliier and screen invertion controlling with physical buttons
+// Listen enables frequency, multipliier and screen invertion controlling with physical buttons.
 func Listen(wg *sync.WaitGroup, ctx context.Context, cfg *config.SafeConfig) {
 	buttons := [4]rpio.Pin{buttonDown, buttonUp, buttonMultiplier, buttonInvert}
 	defer wg.Done()
