+++ pkg/leds/blue.go
@@ -7,12 +7,12 @@ import (
 	"time"
 )
 
-// BlueLedEnabled switches blue LED on or off, if BlueLedLoop is running
+// BlueLedEnabled switches blue LED on or off, if BlueLedLoop is running.
 var BlueLedEnabled = false
 var blueLedSleepInterval = 500 * time.Millisecond
 var blueLed = rpio.Pin(7)
 
-// BlueLedLoop handles blue LED activity
+// BlueLedLoop handles blue LED activity.
 func BlueLedLoop(wg *sync.WaitGroup, ctx context.Context) {
 	defer wg.Done()
 	for {
