+++ pkg/leds/leds.go
@@ -5,14 +5,14 @@ import (
 	"time"
 )
 
-// Blink turns on selected LED and after some time turns it off
+// Blink turns on selected LED and after some time turns it off.
 func Blink(led rpio.Pin, duration time.Duration) {
 	led.High()
 	time.Sleep(duration)
 	led.Low()
 }
 
-// Init sets up all needed LEDs
+// Init sets up all needed LEDs.
 func Init() {
 	greenLed1.Output()
 	greenLed2.Output()
