+++ pkg/leds/quadgreens.go
@@ -16,7 +16,7 @@ var (
 
 var greensSleepInterval = 500 * time.Millisecond
 
-// QuadGreensLoop handles quad green LEDs activity - just blinking in loop in circle
+// QuadGreensLoop handles quad green LEDs activity - just blinking in loop in circle.
 func QuadGreensLoop(wg *sync.WaitGroup, ctx context.Context) {
 	defer wg.Done()
 	greenLed4.High()
