+++ pkg/leds/yellow.go
@@ -7,7 +7,7 @@ import (
 
 var yellowLed = rpio.Pin(5)
 
-// YellowBlink blinks with yellow LED (Blink wrapper)
+// YellowBlink blinks with yellow LED (Blink wrapper).
 func YellowBlink() {
 	Blink(yellowLed, 2*time.Second)
 }
