+++ pkg/logs/logs.go
@@ -14,6 +14,10 @@ func FmRadStrInfo(info interface{}) {
 	FmRadStr.Info().Send("%v", info)
 }
 
+func FmRadStrWarn(err interface{}) {
+	FmRadStr.Error().Send("%v", err)
+}
+
 func FmRadStrFatal(warn interface{}) {
 	FmRadStr.Fatal().Send("%v", warn)
 }
@@ -26,19 +30,6 @@ func PiFmAdvInfo(info interface{}) {
 	PiFmAdv.Info().Send("%v", info)
 }
 
-func PiFmAdvWarn(warn interface{}) {
-	PiFmAdv.Warn().Send("%v", warn)
-}
-
 func PiFmAdvError(err interface{}) {
 	PiFmAdv.Error().Send("%v", err)
-}
-
-// PiFmAdvHandleError uses PiFmAdvError and returns true, if app should return
-func PiFmAdvHandleError(err error) bool {
-	if err != nil {
-		PiFmAdvError(err)
-		return true
-	}
-	return false
 }
