+++ pkg/tools/tools.go
@@ -4,32 +4,12 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/MrBoombastic/FmRadioStreamer/pkg/config"
-	"github.com/MrBoombastic/FmRadioStreamer/pkg/logs"
-	"github.com/arduino/go-apt-client"
 	"github.com/stianeikeland/go-rpio/v4"
 	"io"
-	"net"
 	"net/http"
 	urltool "net/url"
-	"os"
-	"os/exec"
-	"strconv"
 	"time"
 )
-
-// LocalIP stores local IP of the device
-var LocalIP net.IP
-
-// RefreshLocalIP fetches current local IP and saves it to LocalIP variable
-func RefreshLocalIP() {
-	conn, err := net.Dial("udp", "8.8.8.8:80") // It will not actually connect
-	if err != nil {
-		logs.PiFmAdvWarn("Failed to get local IP address! Falling back to localhost...")
-		LocalIP = net.ParseIP("127.0.0.1")
-	}
-	defer conn.Close()
-	LocalIP = conn.LocalAddr().(*net.UDPAddr).IP
-}
 
 // InitGPIO opens connection for LEDs and buttons
 func InitGPIO() error {
@@ -78,58 +58,6 @@ func SearchYouTube(query string, apikey string) (YouTubeAPIResult, error) {
 	}
 
 	return result, nil
-}
-
-// CheckRoot checks if user has root permissions
-func CheckRoot() bool {
-	return os.Geteuid() == 0
-}
-
-func CheckLibsndfileVersion() (float64, error) {
-	lib, err := apt.Search("libsndfile1-dev")
-	if err != nil || len(lib) < 1 {
-		return 0, err
-	}
-	ver := lib[0].Version
-	floatVer, err := strconv.ParseFloat(ver[0:3], 64)
-	if err != nil {
-		return 0, err
-	}
-	return floatVer, nil
-}
-
-func TextToFile(text string, filename string) error {
-	file, err := os.Create(filename)
-	if err != nil {
-		return err
-	}
-	defer func(file *os.File) {
-		err := file.Close()
-		if err != nil {
-			logs.FmRadStrError(err)
-			return
-		}
-	}(file)
-
-	_, err = file.WriteString(text)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func ExecCommand(name string, verbose bool, args ...string) error {
-	cmd := exec.Command(name, args...)
-	// Redirecting stdout and stdin from child process to master process
-	if verbose {
-		cmd.Stdout = os.Stdout
-	}
-	cmd.Stderr = os.Stderr
-	err := cmd.Run()
-	if err != nil {
-		return err
-	}
-	return nil
 }
 
 // ConfigToMap converts given config to map. Needs mutex locked!
